I.matr {enhancer}	R Documentation

Identity relationship matrix

Description

Shortcut to get an identity relationship matrix.

Usage

I.matr(x)

Arguments

x
Vector of values to identify the unique values and form an iende.

Details

Nothing but a shortcut to avoid few lines of code.

Value

If return.imputed = FALSE, the n \times n identity matrix is returned.

$I

the I matrix

References

Giovanny Covarrubias-Pazaran (2024). lme4breeding: enabling genetic evaluation in the age of genomic data. To be submitted to Bioinformatics.

Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48.

Examples

## random population of 200 lines with 1000 markers

[Package enhancer version 1.1.0 Index]