bathy.colors {enhancer}	R Documentation

Generate a sequence of colors for plotting bathymetric data.

Description

bathy.colors(n) generates a sequence of n colors along a linear scale from light grey to pure blue.

Usage

  bathy.colors(n, alpha = 1)

Arguments

n
The number of colors to return.

alpha
Alpha values to be passed to rgb().

Value

A vector of blue scale colors.

Examples

{
# Plot a colorbar using bathy.colors
image(matrix(seq(100), 100), col=bathy.colors(100))
}

[Package enhancer version 1.1.0 Index]