enhancer-package {enhancer}	R Documentation

Mixed Effect Model Enhancing Functions

Description

enhancer is a collection of functions useful in the fitting of linear mixed effect models and general data manipulation. As more of my packages use the same special functions I decided to have the special functions as its own package to facilitate the maintenance.

Functions for genetic analysis

The package provides kernels to the estimate additive (A.matr) relationship matrix for diploid and polyploid organisms. A good converter from letter code to numeric format is implemented in the function atcg1234 and atcg1234BackTransform, which supports higher ploidy levels than diploid. Additional functions for genetic analysis have been included such as build a genotypic hybrid marker matrix (build.HMM). If you need to use pedigree you need to convert your pedigree into a relationship matrix (use the 'getA' function from the pedigreemm package).

Also the LD.decay and the neMarker functions can be used to calculate linkage disequilibrium and effective population size.

Functions for trial analysis

Recently, spatial modeling has been added added to enhancer using the two-dimensional spline tps function. Based on bbasis

Functions for special mixed models

Available are add.diallel.vars, leg, overlay, rrm, redmm, simGECorMat function.

Functions for data manipulation and others

Available are adiag1, imputev, propMissing replaceValues, stan, logspace function.

Simple coloring functions available are jet.colors, bathy.colors, transp

Keeping enhancer updated

The enhancer package is updated on CRAN every 4-months due to CRAN policies but you can find the latest source at https://github.com/covaruber/enhancer. This can be easily installed typing the following in the R console:

library(devtools)

install_github("covaruber/enhancer")

This is recommended if you reported a bug, was fixed and was immediately pushed to GitHub but not in CRAN until the next update.

Bug report and contact

If you have any questions or suggestions please post it in https://stackoverflow.com or https://stats.stackexchange.com

I'll be glad to help or answer any question. I have spent a valuable amount of time developing this package. Please cite this package in your publication. Type 'citation("enhancer")' to know how to cite it.

Author(s)

Giovanny Covarrubias-Pazaran

[Package enhancer version 1.1.0 Index]