jet.colors {enhancer}	R Documentation

Generate a sequence of colors alog the jet colormap.

Description

jet.colors(n) generates a sequence of n colors from dark blue to cyan to yellow to dark red. It is similar to the default color schemes in Python's matplotlib or MATLAB.

Usage

  jet.colors(n, alpha = 1)

Arguments

n
The number of colors to return.

alpha
The transparency value of the colors. See ?rgb for details.

Value

A vector of colors along the jet colorramp.

Examples

{
# Plot a colorbar with jet.colors
image(matrix(seq(100), 100), col=jet.colors(100))
}

[Package enhancer version 1.1.0 Index]