logspace {enhancer}	R Documentation

Decreasing logarithmic trend

Description

logspace creates a vector with decreasing logaritmic trend.

Usage

  logspace(x, p=2)

Arguments

x
sequence of values to pass through the function.

p
power to be applied to the values.

Value

$res

a vector of length n with logarithmic decrease trend.

Author(s)

Giovanny Covarrubias-Pazaran

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Examples

plot(logspace(1:100,p=1))
plot(logspace(1:100,p=2))
plot(logspace(1:100,p=3))

[Package enhancer version 1.1.0 Index]