propMissing {enhancer}	R Documentation

Proportion of missing data

Description

propMissing quick calculation of the proportion of missing data in a vector.

Usage

  propMissing(x)

Arguments

x
vector of observations.

Value

$res

a numeric value with the proportion of missing data.

Author(s)

Giovanny Covarrubias-Pazaran

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommec. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Examples

x <- c(1:10, NA)
propMissing(x)

[Package enhancer version 1.1.0 Index]