simGECorMat {enhancer}	R Documentation

Create a GE correlation matrix for simulation purposes.

Description

Makes a simple correlation matrix based on the number of environments and megaenvironments desired.

Usage

simGECorMat(nEnv,nMegaEnv,mu=0.7, v=0.2, mu2=0, v2=0.3)

Arguments

nEnv
Number of environments to simulate. Needs to be divisible by the nMegaEnv argument.

nMegaEnv
Number of megaenvironments to simulate.

mu
Mean value of the genetic correlation within megaenvironments.

v
variance in the genetic correlation within megaenvironments.

mu2
Mean value of the genetic correlation between megaenvironments.

v2
variance in the genetic correlation between megaenvironments.

Details

Simple simulation of a correlation matrix for environments and megaenvironments.

Value

G the correlation matrix

$G

the correlation matrix

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Examples

simGECorMat(9,3)

[Package enhancer version 1.1.0 Index]