stan {enhancer}	R Documentation

Standardize a vector of values in range 0 to 1

Description

Simple function to map a vector of values to the range of 0 and 1 values to have a better behavior of the algorithm.

Usage

stan(x, lb=0, ub=1)

Arguments

x
A vector of numeric values.

lb
Lower bound value to map the x values.

ub
Upper bound value to map the x values.

Details

Simple function to map a vector of values to the range of 0 and 1 values to have a better behavior of the algorithm.

Value

$res

new values in range 0 to 1

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Examples

x <- rnorm(20, 10, 3);x
stan(x)

[Package enhancer version 1.1.0 Index]